+++ pkg/test/main/main.go
@@ -41,6 +41,33 @@ import (
 	resourcev3 "github.com/envoyproxy/go-control-plane/pkg/test/resource/v3"
 )
 
+// Flags type
+type Flags struct {
+	Debug bool
+
+	Port         uint
+	GatewayPort  uint
+	UpstreamPort uint
+	BasePort     uint
+	AlsPort      uint
+
+	Delay    time.Duration
+	Requests int
+	Updates  int
+
+	Mode          string
+	Clusters      int
+	HTTPListeners int
+	TCPListeners  int
+	Runtimes      int
+	TLS           bool
+
+	NodeID string
+
+	V2 bool
+	V3 bool
+}
+
 var flags Flags
 
 func init() {
